Add implementation of OfferCrud. Remove exceed service from DI
Merge branch 'release/0.1.0' of github.com:Anakael/Treasury-Department into feature/core/friendservice

# Conflicts:
#	TreasuryDepartment/Models/BlackList.cs
#	TreasuryDepartment/Models/Friend.cs
#	TreasuryDepartment/Models/TreasuryDepartmentContext.cs
